Given the tickers, statement, year and period returns all the information from the Intrinio API fundamental reported financials for that time and those tickers as a dataframe

Given the tickers, statement, year and period returns all the information from the Intrinio API fundamental reported financials for that time and those tickers as a dataframe

gather_financial_statement_company_compare(
  api_key,
  ticker,
  statement,
  year,
  period
)

Arguments

api_key
character API key (sandbox or production) from Intrinio

ticker
character vector of the ticker symbols you would like to study

statement
character the statement that you want to study options: 'income_statement', 'cash_flow_statement', 'balance_sheet_statement'

year
character the year you want the information from

period
character the period you want the information from

Value

a dataframe that contains information about the given statement for the given tickers at the given time

Examples

gather_financial_statement_company_compare('OjhlMjhjNTBmY2IyMWJiMWE0MTExYjQwNWZmZTVkZWM1',
c('AAPL', 'CSCO'), 'income_statement', '2019', 'Q1')

#> Warning: Column `name` joining factors with different levels, coercing to character vector

#>                               name          value.x          value.y
#> 1                           ticker             AAPL             CSCO
#> 2                        statement income_statement income_statement
#> 3                             year             2019             2019
#> 4                           period               Q1               Q1
#> 5                 operatingrevenue       -8.431e+10      -1.3072e+10
#> 6                     totalrevenue       -8.431e+10      -1.3072e+10
#> 7           operatingcostofrevenue       5.2279e+10        4.926e+09
#> 8               totalcostofrevenue       5.2279e+10        4.926e+09
#> 9                 totalgrossprofit      -3.2031e+10       -8.146e+09
#> 10                      sgaexpense        4.783e+09         2.11e+08
#> 11                       rdexpense        3.902e+09        1.608e+09
#> 12          totaloperatingexpenses        8.685e+09        4.341e+09
#> 13            totaloperatingincome      -2.3346e+10       -3.805e+09
#> 14                     otherincome         -5.6e+08          1.9e+07
#> 15                totalotherincome         -5.6e+08        -1.04e+08
#> 16               totalpretaxincome      -2.3906e+10       -3.909e+09
#> 17                incometaxexpense        3.941e+09          3.6e+08
#> 18             netincomecontinuing      -1.9965e+10       -3.549e+09
#> 19                       netincome      -1.9965e+10       -3.549e+09
#> 20               netincometocommon      -1.9965e+10       -3.549e+09
#> 21        weightedavebasicsharesos      -4735820000       -4.565e+09
#> 22                        basiceps            -4.22            -0.78
#> 23      weightedavedilutedsharesos      -4773252000       -4.614e+09
#> 24                      dilutedeps            -4.18            -0.77
#> 25 weightedavebasicdilutedsharesos       -4.731e+09        -4.55e+09
#> 26                 basicdilutedeps            -4.22            -0.78
#> 27           cashdividendspershare            -0.73            -0.33
#> 28                marketingexpense             <NA>         2.41e+09
#> 29             amortizationexpense             <NA>          3.4e+07
#> 30             restructuringcharge             <NA>          7.8e+07
#> 31            totalinterestexpense             <NA>         2.21e+08
#> 32             totalinterestincome             <NA>        -3.44e+08