Given the ticker, statement, year(s) and period(s) return the complete available financial information from the Intrinio API stock data

Given the ticker, statement, year(s) and period(s) return the complete available financial information from the Intrinio API stock data

gather_financial_statement_time_series(
  api_key,
  ticker,
  statement,
  year,
  period
)

Arguments

api_key
character (sandbox or production) from Intrinio

ticker
character the ticker symbol you would like to get information for

statement
character that represents the financial statement that you want to study options: 'income_statement', 'cash_flow_statement', 'balance_sheet_statement'

year
character vector of the year(s) you want the information for

period
character vector of the period(s) you want the information for

Value

dataframe containing information about the given statement for the given ticker at the given times

Examples

gather_financial_statement_time_series('OjhlMjhjNTBmY2IyMWJiMWE0MTExYjQwNWZmZTVkZWM1',
'CVX', 'income_statement', c('2017','2018'), c('Q1','Q3'))

#>                                 type        fin_value       fin_value.
#> 1                             ticker              CVX              CVX
#> 2                          statement income_statement income_statement
#> 3                               year             2017             2017
#> 4                             period               Q1               Q3
#> 5                   operatingrevenue      -3.3421e+10      -3.6205e+10
#> 6                       totalrevenue      -3.3421e+10      -3.6205e+10
#> 7             operatingcostofrevenue       1.7506e+10       1.8776e+10
#> 8                 totalcostofrevenue       1.7506e+10       1.8776e+10
#> 9                   totalgrossprofit      -1.5915e+10      -1.7429e+10
#> 10                        sgaexpense        5.396e+09        5.956e+09
#> 11                explorationexpense         1.44e+08         2.39e+08
#> 12               depreciationexpense        4.194e+09        5.109e+09
#> 13            otheroperatingexpenses        3.001e+09        3.432e+09
#> 14            totaloperatingexpenses       1.2735e+10       1.4736e+10
#> 15              totaloperatingincome        -3.18e+09       -2.693e+09
#> 16              totalinterestexpense          5.1e+07          3.5e+07
#> 17                  totalotherincome          5.1e+07          3.5e+07
#> 18                 totalpretaxincome       -3.129e+09       -2.658e+09
#> 19                  incometaxexpense          4.3e+08         6.72e+08
#> 20               netincomecontinuing       -2.699e+09       -1.986e+09
#> 21                         netincome       -2.699e+09       -1.986e+09
#> 22 netincometononcontrollinginterest          1.7e+07          3.4e+07
#> 23                 netincometocommon       -2.682e+09       -1.952e+09
#> 24          weightedavebasicsharesos      -1879372000      -1882650000
#> 25                          basiceps            -1.43            -1.03
#> 26        weightedavedilutedsharesos      -1895393000      -1895879000
#> 27                        dilutedeps            -1.41            -1.03
#> 28   weightedavebasicdilutedsharesos      -1875500000      -1895100000
#> 29                   basicdilutedeps            -1.43            -1.03
#> 30             cashdividendspershare            -1.08            -1.08
#>         fin_value..     fin_value...
#> 1               CVX              CVX
#> 2  income_statement income_statement
#> 3              2018             2018
#> 4                Q1               Q3
#> 5       -3.7764e+10      -4.3987e+10
#> 6       -3.7764e+10      -4.3987e+10
#> 7        2.1233e+10       2.4681e+10
#> 8        2.1233e+10       2.4681e+10
#> 9       -1.6531e+10      -1.9306e+10
#> 10        5.424e+09        6.003e+09
#> 11         1.58e+08         6.25e+08
#> 12        4.289e+09         5.38e+09
#> 13        1.428e+09        1.417e+09
#> 14       1.1299e+10       1.3425e+10
#> 15       -5.232e+09       -5.881e+09
#> 16         1.59e+08         1.82e+08
#> 17         1.59e+08         1.82e+08
#> 18       -5.073e+09       -5.699e+09
#> 19         1.41e+09        1.643e+09
#> 20       -3.663e+09       -4.056e+09
#> 21       -3.659e+09       -4.056e+09
#> 22          2.1e+07            9e+06
#> 23       -3.638e+09       -4.047e+09
#> 24      -1895990000      -1900717000
#> 25            -1.92            -2.13
#> 26      -1913218000      -1917474000
#> 27             -1.9            -2.11
#> 28      -1894800000         -1.9e+09
#> 29            -1.92            -2.13
#> 30            -1.12            -1.12